T316 Rename directory to be inline with others
